+++ packages/monaco/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.4.16](http://github.com///compare/@codecademy/monaco@0.4.15...@codecademy/monaco@0.4.16) (2021-10-05)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
 ### [0.4.15](http://github.com///compare/@codecademy/monaco@0.4.14...@codecademy/monaco@0.4.15) (2021-10-01)
 
 **Note:** Version bump only for package @codecademy/monaco